refactor(client-3d): split NetworkManager handlers + remove network→scene dep
Phase 3C+3D:
- Split NetworkManager.ts 1042→664 LOC. wireRoomListeners was 350
  tangled LOC mixing 7+ domains; now delegates to focused handlers:
  players, chat, music, djQueue, jukebox, jump, lifecycle.
- Introduced network/events.ts pub/sub for network→scene signals.
  The jump handler now emits PlayerJump events instead of importing
  triggerRemoteJump/addRipple from scene/. GameScene subscribes via
  a useRemoteJumpBridge effect, eliminating the network→scene
  circular-layer dependency flagged in the audit.
