+++ client-3d/src/network/NetworkManager.ts
@@ -1,23 +1,22 @@
-import { Client, Room, getStateCallbacks } from '@colyseus/sdk'
-import type { RoomState, Player } from '@club-mutant/types/RoomState'
+import { Client, Room } from '@colyseus/sdk'
+import type { RoomState } from '@club-mutant/types/RoomState'
 import { Message } from '@club-mutant/types/Messages'
 import { RoomType } from '@club-mutant/types/Rooms'
 import type { RoomListEntry } from '../stores/gameStore'
 import { useAuthStore } from '../stores/authStore'
 import { getValidToken } from './nakamaClient'
 
-import { useGameStore, setPlayerPosition, getPlayerPosition } from '../stores/gameStore'
+import { useGameStore } from '../stores/gameStore'
 import { useChatStore } from '../stores/chatStore'
-import { useMusicStore } from '../stores/musicStore'
 import { useBoothStore } from '../stores/boothStore'
-import { useJukeboxStore } from '../stores/jukeboxStore'
-import { useUIStore } from '../stores/uiStore'
-import { useToastStore } from '../stores/toastStore'
-import { playNpcTtsAudio } from '../audio/npcTtsPlayer'
-import type { JukeboxItemDto } from '@club-mutant/types/Dtos'
-import { triggerRemoteJump } from '../scene/PlayerEntity'
-import { addRipple } from '../scene/TrampolineRipples'
 import { TimeSync } from './TimeSync'
+import { wirePlayerHandlers } from './messages/playerHandlers'
+import { wireChatHandlers } from './messages/chatHandlers'
+import { wireMusicHandlers } from './messages/musicHandlers'
+import { wireDJQueueHandlers } from './messages/djQueueHandlers'
+import { wireJukeboxHandlers } from './messages/jukeboxHandlers'
+import { wireJumpHandlers } from './messages/jumpHandlers'
+import { wireLifecycleHandlers } from './messages/lifecycleHandlers'
 
 const PLAYER_ID_KEY = 'club-mutant-3d:player-id'
 
@@ -403,395 +402,18 @@ export class NetworkManager {
   private wireRoomListeners() {
     if (!this.room) return
 
-    const $ = getStateCallbacks(this.room)
-
-    // Colyseus 0.17 pattern: wrap state, then access collections on the proxy
-    const stateProxy = $(this.room.state) as any
-    const playersProxy = stateProxy.players
-
-    // Clamp server positions to 3D room bounds (ROOM_SIZE=12, WORLD_SCALE=0.01 → ±550 server px)
-    const ROOM_MAX = 550
-    const clampPos = (v: number) => Math.max(-ROOM_MAX, Math.min(ROOM_MAX, v))
-
-    // Track recently-added remote players so we can skip initial listen fires
-    // that carry the server's legacy 2D default position (705, 500).
-    const freshRemotes = new Set<string>()
-
-      // Player add/remove/change
-      playersProxy.onAdd((player: Player, sessionId: string) => {
-        const gameStore = useGameStore.getState()
-        const chatStore = useChatStore.getState()
-
-        const cx = clampPos(player.x)
-        const cy = clampPos(player.y)
-
-        console.log(
-          `[network] onAdd ${sessionId} textureId=${player.textureId} name=${player.name} pos=(${player.x},${player.y})→(${cx},${cy})`
-        )
-
-        const isLocal = sessionId === this.room?.sessionId
-        const localSessionId = this.room?.sessionId
-        
-        const playerProxy = $(player) as any
-
-        playerProxy.listen('x', (value: number) => {
-          if (sessionId === localSessionId) return
-          const pos = getPlayerPosition(sessionId)
-          if (pos) pos.x = clampPos(value)
-        })
-
-        playerProxy.listen('y', (value: number) => {
-          if (sessionId === localSessionId) return
-          const pos = getPlayerPosition(sessionId)
-          if (pos) pos.y = clampPos(value)
-        })
-
-        setPlayerPosition(sessionId, cx, cy)
-
-      gameStore.addPlayer(sessionId, {
-        sessionId,
-        name: player.name,
-        textureId: player.textureId,
-        animId: player.animId,
-        scale: player.scale,
-        isNpc: player.isNpc ?? false,
-        npcCharacterPath: player.npcCharacterPath ?? '',
-        npcAnimState: player.npcAnimState ?? '',
-        nakamaId: player.nakamaId ?? '',
-      })
-
-      if (isLocal) {
-        gameStore.setLocalPosition(0, 0)
-        // Tell the server we're at (0,0) so it doesn't echo back the 2D default
-        this.sendPosition(0, 0, 'idle')
-      }
-
-      playerProxy.listen('name', (value: string) => {
-        useGameStore.getState().updatePlayer(sessionId, { name: value })
-      })
-
-      playerProxy.listen('animId', (value: number) => {
-        useGameStore.getState().updatePlayer(sessionId, { animId: value })
-      })
-
-      playerProxy.listen('textureId', (value: number) => {
-        console.log(`[network] textureId listen ${sessionId} -> ${value}`)
-        useGameStore.getState().updatePlayer(sessionId, { textureId: value })
-      })
-
-      playerProxy.listen('scale', (value: number) => {
-        useGameStore.getState().updatePlayer(sessionId, { scale: value })
-      })
-
-      playerProxy.listen('npcAnimState', (value: string) => {
-        useGameStore.getState().updatePlayer(sessionId, { npcAnimState: value })
-      })
-
-      // Don't show "joined" message for NPC players — they're always present
-      if (player.name && !player.isNpc) {
-        chatStore.addMessage({
-          id: crypto.randomUUID(),
-          author: 'system',
-          content: `${player.name} joined`,
-          createdAt: Date.now(),
-        })
-      }
-    }, true) // true = trigger for existing items
-
-    playersProxy.onRemove((_player: Player, sessionId: string) => {
-      const gameStore = useGameStore.getState()
-      const chatStore = useChatStore.getState()
-
-      const existing = gameStore.players.get(sessionId)
-      const name = existing?.name ?? sessionId
-
-      gameStore.removePlayer(sessionId)
-
-      // Don't show "left" message for NPC players
-      if (!existing?.isNpc) {
-        chatStore.addMessage({
-          id: crypto.randomUUID(),
-          author: 'system',
-          content: `${name} left`,
-          createdAt: Date.now(),
-        })
-      }
+    wirePlayerHandlers(this.room, () => {
+      // Tell the server we're at (0,0) so it doesn't echo back the 2D default
+      this.sendPosition(0, 0, 'idle')
     })
-
-    // Chat messages (from other players)
-    this.room.onMessage(Message.ADD_CHAT_MESSAGE, (data: { clientId: string; content: string; imageUrl?: string }) => {
-      const player = useGameStore.getState().players.get(data.clientId)
-      const chatStore = useChatStore.getState()
-
-      chatStore.addMessage({
-        id: crypto.randomUUID(),
-        author: player?.name ?? data.clientId,
-        content: data.content,
-        imageUrl: data.imageUrl || undefined,
-        createdAt: Date.now(),
-      })
-
-      // Show in-world chat bubble
-      if (data.clientId) {
-        chatStore.setBubble(data.clientId, data.content, data.imageUrl || undefined)
-      }
-    })
-
-    // NPC TTS audio (base64 WAV from SAPI4)
-    this.room.onMessage(Message.NPC_TTS_AUDIO, (data: { audioBase64: string; durationMs: number }) => {
-      if (data.audioBase64) {
-        playNpcTtsAudio(data.audioBase64)
-      }
-    })
-
-    // Chat history (bulk load on join)
-    this.room.onMessage(Message.CHAT_HISTORY, (messages: Array<{ author: string; content: string; imageUrl: string; createdAt: number }>) => {
-      const chatStore = useChatStore.getState()
-      chatStore.setMessages(
-        messages.map((m) => ({
-          id: crypto.randomUUID(),
-          author: m.author,
-          content: m.content,
-          imageUrl: m.imageUrl || undefined,
-          createdAt: m.createdAt,
-        }))
-      )
-    })
-
-    // Music stream messages
-    this.room.onMessage(
-      Message.START_MUSIC_STREAM,
-      (data: { musicStream: any; offset: number }) => {
-        const ms = data.musicStream
-        if (!ms) return
-
-        // Skip ambient background streams — no DJ is playing
-        if (ms.isAmbient) return
-
-        // Convert server startTime to client time using clock sync
-        const clientStartTime = this._timeSync?.ready
-          ? this._timeSync.toClientTime(ms.startTime ?? 0)
-          : (ms.startTime ?? 0)
-
-        useMusicStore.getState().setStream({
-          currentLink: ms.currentLink ?? null,
-          currentTitle: ms.currentTitle ?? null,
-          currentDjName: ms.currentDj?.name ?? null,
-          startTime: clientStartTime,
-          duration: ms.duration ?? 0,
-          isPlaying: true,
-          streamId: ms.streamId ?? 0,
-        })
-      }
-    )
-
-    this.room.onMessage(Message.STOP_MUSIC_STREAM, () => {
-      useMusicStore.getState().clearStream()
-    })
-
-    // Periodic drift correction — server sends streamId + startTime every 5s
-    this.room.onMessage(
-      Message.MUSIC_STREAM_TICK,
-      (data: { streamId: number; startTime: number; serverNowMs: number }) => {
-        const store = useMusicStore.getState()
-
-        // Ignore ticks for a different stream
-        if (!store.stream.isPlaying || store.stream.streamId !== data.streamId) return
-
-        // Recompute client-local startTime from this tick's authoritative server time
-        const clientStartTime = this._timeSync?.ready
-          ? this._timeSync.toClientTime(data.startTime)
-          : data.startTime
-
-        const currentClientStart = store.stream.startTime
-        const drift = Math.abs(clientStartTime - currentClientStart)
-
-        // Only correct if drift > 2 seconds to avoid micro-jitter
-        if (drift > 2000) {
-          console.log(`[TimeSync] Drift correction: ${drift}ms, resyncing startTime`)
-
-          store.setStream({ startTime: clientStartTime })
-        }
-      }
-    )
-
-    // Helper to sync DJ queue from schema state into boothStore
-    const syncDJQueueFromSchema = () => {
-      const rs = this.room?.state as any
-      if (!rs) return
-
-      const dq = rs.djQueue
-      const entries: import('../stores/boothStore').DJQueueEntry[] = dq
-        ? Array.from(dq as Iterable<any>).map((e: any) => ({
-            sessionId: e.sessionId as string,
-            name: e.name as string,
-            position: (e.queuePosition ?? 0) as number,
-            slotIndex: (e.slotIndex ?? 0) as number,
-          }))
-        : []
-
-      const booth = useBoothStore.getState()
-      booth.setDJQueue(entries, rs.currentDjSessionId ?? null)
-
-      const myId = this.room?.sessionId
-      booth.setIsInQueue(entries.some((e) => e.sessionId === myId))
-    }
-
-    // Schema callbacks — sole mechanism for DJ queue sync.
-    // Fires on late-join (initial state delivery) AND on every live mutation
-    // (join/leave/skip/rotation). No separate message handler needed.
-    try {
-      const djQueueProxy = stateProxy.djQueue
-      djQueueProxy.onAdd(() => syncDJQueueFromSchema())
-      djQueueProxy.onRemove(() => syncDJQueueFromSchema())
-      stateProxy.listen('currentDjSessionId', () => syncDJQueueFromSchema())
-    } catch (err) {
-      console.warn('[network] Schema callbacks for djQueue failed:', err)
-    }
-
-    // Jukebox playlist schema callbacks — syncs shared playlist to jukeboxStore.
-    // Fires on late-join (initial state) and every live mutation (add/remove/splice).
-    try {
-      const jukeboxPlaylistProxy = stateProxy.jukeboxPlaylist
-
-      const syncJukeboxPlaylist = () => {
-        const rs = this.room?.state as any
-        if (!rs?.jukeboxPlaylist) return
-
-        const items: JukeboxItemDto[] = Array.from(
-          rs.jukeboxPlaylist as Iterable<any>
-        ).map((item: any) => ({
-          id: item.id as string,
-          title: item.title as string,
-          link: item.link as string,
-          duration: (item.duration ?? 0) as number,
-          addedBySessionId: item.addedBySessionId as string,
-          addedByName: item.addedByName as string,
-          addedAtMs: (item.addedAtMs ?? 0) as number,
-        }))
-
-        useJukeboxStore.getState().setPlaylist(items)
-      }
-
-      jukeboxPlaylistProxy.onAdd(() => syncJukeboxPlaylist())
-      jukeboxPlaylistProxy.onRemove(() => syncJukeboxPlaylist())
-    } catch (err) {
-      console.warn('[network] Schema callbacks for jukeboxPlaylist failed:', err)
-    }
-
-    // Jukebox occupant sync — track who's using the jukebox
-    try {
-      const syncJukeboxOccupant = () => {
-        const rs = this.room?.state as any
-        const uid = rs?.jukeboxUserId ?? ''
-        const uname = rs?.jukeboxUserName ?? ''
-        useJukeboxStore
-          .getState()
-          .setOccupant(uid || null, uname || null)
-
-        // Auto-open panel if we just became the occupant
-        if (uid === useGameStore.getState().mySessionId) {
-          useUIStore.getState().setDjQueueOpen(true)
-        }
-      }
-
-      stateProxy.listen('jukeboxUserId', syncJukeboxOccupant)
-      stateProxy.listen('jukeboxUserName', syncJukeboxOccupant)
-
-      // Late-join: sync initial occupant state
-      syncJukeboxOccupant()
-    } catch (err) {
-      console.warn('[network] Schema callbacks for jukebox occupant failed:', err)
-    }
-
-    // Jukebox busy rejection — server tells us someone else is using it
-    this.room.onMessage('jukebox_busy', (data: { name: string }) => {
-      useToastStore.getState().addToast(`${data.name} is using the jukebox`)
-    })
-
-    // Late-join: sync music state AFTER TimeSync is ready (correct seek offset)
-    if (this._timeSync) {
-      this._timeSync.onReady(() => {
-        const rs = this.room?.state as any
-        const ms = rs?.musicStream
-        if (!ms || ms.status !== 'playing' || !ms.currentLink || ms.isAmbient) return
-
-        // Skip if START_MUSIC_STREAM message already set this stream
-        const store = useMusicStore.getState()
-        if (store.stream.isPlaying && store.stream.streamId === (ms.streamId ?? 0)) return
-
-        const clientStartTime = this._timeSync!.toClientTime(ms.startTime ?? 0)
-
-        store.setStream({
-          currentLink: ms.currentLink,
-          currentTitle: ms.currentTitle ?? null,
-          currentDjName: ms.currentDj?.name ?? null,
-          startTime: clientStartTime,
-          duration: ms.duration ?? 0,
-          isPlaying: true,
-          streamId: ms.streamId ?? 0,
-        })
-      })
-    }
-
-    // Per-player queue playlist updates
-    this.room.onMessage(Message.ROOM_QUEUE_PLAYLIST_UPDATED, (payload: { items: any[] }) => {
-      useBoothStore.getState().setQueuePlaylist(
-        payload.items.map((item) => ({
-          id: item.id,
-          title: item.title,
-          link: item.link,
-          duration: item.duration ?? 0,
-          played: item.played ?? false,
-        }))
-      )
-    })
-
-    // Trampoline jump from other players
-    this.room.onMessage(Message.PLAYER_JUMP, (data: { sessionId: string }) => {
-      if (!data.sessionId) return
-
-      triggerRemoteJump(data.sessionId)
-
-      // Create the floor ripple immediately using the player's current store position.
-      // This avoids timing issues where the PlayerEntity's useFrame hasn't run yet.
-      const WORLD_SCALE = 0.01
-      const TAKEOFF_RIPPLE_AMP = 0.08
-      const pos = getPlayerPosition(data.sessionId)
-
-      if (pos) {
-        const wx = pos.x * WORLD_SCALE
-        const wz = -pos.y * WORLD_SCALE
-        addRipple(wx, wz, TAKEOFF_RIPPLE_AMP)
-      }
-    })
-
-    // Reconnection: connection dropped unexpectedly
-    this.room.onDrop((code: number, reason?: string) => {
-      console.log(`[network] Connection dropped! code=${code} reason=${reason}`)
-      useGameStore.getState().setConnectionStatus('reconnecting')
-    })
-
-    // Reconnection: successfully reconnected
-    this.room.onReconnect(() => {
-      console.log('[network] Reconnected successfully!')
-      useGameStore.getState().setConnectionStatus('connected')
-
-      // Re-sync TimeSync after reconnection
-      this._timeSync?.start()
-    })
-
-    // Room leave — permanent (either consented or failed to reconnect)
-    this.room.onLeave((code: number) => {
-      console.log('[network] Left room, code:', code)
-      this.releaseSessionLock()
-      useGameStore.getState().setConnectionStatus('disconnected')
-      useMusicStore.getState().clearStream()
-      useBoothStore.getState().setBoothConnected(false)
-      useBoothStore.getState().setIsInQueue(false)
-      useJukeboxStore.getState().clear()
-    })
+    wireChatHandlers(this.room)
+    wireMusicHandlers(this.room, this._timeSync)
+    wireDJQueueHandlers(this.room)
+    wireJukeboxHandlers(this.room)
+    wireJumpHandlers(this.room)
+    wireLifecycleHandlers(this.room, this._timeSync, () => this.releaseSessionLock())
   }
+
 
   // Send chat message (with optional image URL from CDN upload)
   sendChat(content: string, imageUrl?: string) {
